fix code climate warnings
